Commit saved weights and saved model for online
Both for v4 of dataset parse, which I believe was 30-length sequences
drawn across all the data, but without ever having to pad, and yes
reshaped. Double check that though.

Goal is to try to serve this on my mac and then send requests over in
some approximation of realtime - eventually to experiment with online
dataset.