Adding send_live_data.py for sending live data
Starting to play around with the serving element of things to do live
(online) recognition. Just ran an experiment where I ran every
30-element subsequence of a gesture then predicted the outcome based on
the mode of the predictions, and got around 70% accuracy overall.
Interesting. Also of note, it seemed like the most common fuckup was for
the model to identify the gesture as 0 when it wasn't 0, which is
curious because there actually wasn't any 'null' training data for this
model - I just put in an extra element to accomodate the out-of-the-box
labels (1-indexed) of the training set.