Added script to send online dataset data
Downloaded the online recognition dataset - 10 subjects, 8-10 sequences
per with gestures performed itermittently - and derped up a script to
send the data over post request to a tensorflow serving server running
on localhost (in a separate linux command line tab smh). Gonna do a
little clean up now and then start gathering metrics, but on first
glance it didn't look like the online recognition was performing all
that badly. If this is an acceptable level of accuracy then the next
step would be to either rewrite it in pytorch or whatever the currently
hot framework is, or get it working in unity. But probably rewrite since
I don't want to hook up a bunch of unity plumbing that's tensorflow
specific and then have to change it all. Idk we'll see I guess